Move class 'AuthenticationHeaders' to 'Lobby-Client'
The class 'AuthenticationHeaders' is used to create headers that are
specific to game-clients communicating with the lobby. Previously
'AuthenticationHeaders' was in the very generic 'http-client' package,
which is not appropriate since the class is server specific.
