+++ game-app/game-core/src/main/java/games/strategy/engine/lobby/client/login/LobbyLogin.java
@@ -11,9 +11,9 @@ import javax.swing.JFrame;
 import lombok.extern.slf4j.Slf4j;
 import org.triplea.domain.data.ApiKey;
 import org.triplea.domain.data.UserName;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.forgot.password.ForgotPasswordClient;
 import org.triplea.http.client.forgot.password.ForgotPasswordRequest;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.http.client.lobby.login.CreateAccountResponse;
 import org.triplea.http.client.lobby.login.LobbyLoginClient;
 import org.triplea.http.client.lobby.login.LobbyLoginResponse;
