+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/lobby/login/LobbyLoginClient.java
@@ -5,8 +5,8 @@ import java.net.URI;
 import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.HttpClient;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class LobbyLoginClient {
