+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/lobby/moderator/toolbox/banned/user/ToolboxUserBanClient.java
@@ -6,8 +6,8 @@ import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import org.triplea.domain.data.ApiKey;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.HttpClient;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 
 /**
  * Http client object for adding, removing and querying server for user bans. User bans are done by
