+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/lobby/moderator/toolbox/management/ToolboxModeratorManagementClient.java
@@ -8,8 +8,8 @@ import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import org.triplea.domain.data.ApiKey;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.HttpClient;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 
 /**
  * Http client for moderator management. Other than the fetch to display the list of moderators,
