+++ http-clients/lobby-client/src/main/java/org/triplea/http/client/lobby/moderator/toolbox/words/ToolboxBadWordsClient.java
@@ -8,8 +8,8 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.triplea.domain.data.ApiKey;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.HttpClient;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 
 /** Http client class for fetching the list of bad words and adding and removing them. */
 @Slf4j
