+++ http-clients/lobby-client/src/test/java/org/triplea/http/client/lobby/AuthenticationHeadersTest.java
@@ -1,4 +1,4 @@
-package org.triplea.http.client;
+package org.triplea.http.client.lobby;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
