+++ http-clients/lobby-client/src/test/java/org/triplea/http/client/lobby/game/lobby/watcher/GameListingClientTest.java
@@ -13,9 +13,9 @@ import com.github.tomakehurst.wiremock.client.WireMock;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.triplea.domain.data.LobbyGame;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.TestData;
 import org.triplea.http.client.WireMockTest;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.test.common.JsonUtil;
 import ru.lanwen.wiremock.ext.WiremockResolver;
