+++ http-clients/lobby-client/src/test/java/org/triplea/http/client/lobby/moderator/toolbox/log/ToolboxAccessLogClientTest.java
@@ -12,9 +12,9 @@ import java.time.Instant;
 import java.util.List;
 import org.hamcrest.collection.IsCollectionWithSize;
 import org.junit.jupiter.api.Test;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.HttpClientTesting;
 import org.triplea.http.client.WireMockTest;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.test.common.JsonUtil;
 import ru.lanwen.wiremock.ext.WiremockResolver;
