+++ http-clients/lobby-client/src/test/java/org/triplea/http/client/lobby/moderator/toolbox/words/ToolboxBadWordsClientTest.java
@@ -10,8 +10,8 @@ import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.WireMockTest;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.test.common.JsonUtil;
 import ru.lanwen.wiremock.ext.WiremockResolver;
