+++ http-clients/lobby-client/src/test/java/org/triplea/http/client/lobby/user/account/UserAccountClientTest.java
@@ -8,8 +8,8 @@ import static org.triplea.http.client.HttpClientTesting.EXPECTED_API_KEY;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import org.junit.jupiter.api.Test;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.WireMockTest;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.test.common.JsonUtil;
 import ru.lanwen.wiremock.ext.WiremockResolver;
