+++ spitfire-server/dropwizard-server/src/main/java/org/triplea/spitfire/server/controllers/ErrorReportController.java
@@ -9,13 +9,13 @@ import javax.ws.rs.Path;
 import javax.ws.rs.core.Context;
 import lombok.Builder;
 import org.jdbi.v3.core.Jdbi;
-import org.triplea.http.client.AuthenticationHeaders;
 import org.triplea.http.client.error.report.CanUploadErrorReportResponse;
 import org.triplea.http.client.error.report.CanUploadRequest;
 import org.triplea.http.client.error.report.ErrorReportClient;
 import org.triplea.http.client.error.report.ErrorReportRequest;
 import org.triplea.http.client.error.report.ErrorReportResponse;
 import org.triplea.http.client.github.GithubApiClient;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.modules.LobbyModuleConfig;
 import org.triplea.modules.error.reporting.CanUploadErrorReportStrategy;
 import org.triplea.modules.error.reporting.CreateIssueParams;
