+++ spitfire-server/dropwizard-server/src/test/java/org/triplea/spitfire/server/access/authorization/BannedPlayerFilterIntegrationTest.java
@@ -9,7 +9,7 @@ import java.util.Map;
 import lombok.AllArgsConstructor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.triplea.http.client.AuthenticationHeaders;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 import org.triplea.http.client.lobby.login.LobbyLoginClient;
 import org.triplea.http.client.lobby.login.LobbyLoginResponse;
 import org.triplea.spitfire.server.ControllerIntegrationTest;
