+++ spitfire-server/dropwizard-server/src/test/java/org/triplea/spitfire/server/access/authorization/BannedPlayerFilterTest.java
@@ -26,7 +26,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.triplea.db.dao.user.ban.BanLookupRecord;
 import org.triplea.db.dao.user.ban.UserBanDao;
-import org.triplea.http.client.AuthenticationHeaders;
+import org.triplea.http.client.lobby.AuthenticationHeaders;
 
 @SuppressWarnings("SameParameterValue")
 @ExtendWith(MockitoExtension.class)
